[Chore] modify chat page 코드 위치 수정
